Simplify pantherlog jsoniter extension by registering it globally
